applying settings

+++ app/src/main/java/com/example/smiledphoto/data/preferences/Preferences.kt
@@ -5,6 +5,6 @@ import com.example.smiledphoto.data.enumeration.QualityEnum
 
 interface Preferences {
     var quality: QualityEnum
-    var smilingProbabilityThreshold: Int
+    var smilingProbabilityThreshold: Float
     var cameraType: CameraTypeEnum
 }
